requirements.txt
Telethon==1.29.2
python-dotenv==1.0.0